Pending users / UI update
Pending users is fully functional along with the getting location /
mapping it. I have also added in a basic BASIC ui for the pending
requests and my next goal is to take some of the load off of the main
thread with an asynchronous task on the main thread for geocoding.